Dependency management is from now on unnecessary
The system has changed to not have any dependencies except for "load libraries
first, then load the core, then everything else and then sentinel". This is
possible due to the decision that *modules are not allowed to assume the
existence of an outside world before `allModulesLoaded` event fires*. Modules
are becoming less and less aware of how the outside world looks.